Add EEPROM, multiple plans, goHome, serial cmds
For clarity, functions and methods were moved to .ino files.
Which are analogous to .cpp files.

Robot can now be programmed using the serial interface.
The user can sent new plans, store them in one of (currently) 4
persistent slots, and load them later with a command. The slots
are implemented using EEPROM memory so they persist even if
the robot is shut off.

Added an example how can the planning interface be used.